Added Writer.java and optimized formatting
